+++ src/helpers/webview/context_menu.js
@@ -1,9 +1,9 @@
 // Provide context menus (copy, paste, save image, etc...) for right-click interaction.
-// Must not contain certain newer JS syntaxes to allow use inside a webview.
 
-const remote = require('electron').remote;
+import { ipcRenderer, remote } from 'electron';
+import { EVENT_SPELL_ADD_CUSTOM_WORD } from '../../constants';
 
-const Menu = remote.Menu;
+const { Menu } = remote;
 
 const standardMenuTemplate = [
   {
@@ -52,26 +52,35 @@ const textMenuTemplate = [
   }
 ];
 
-const standardInputMenu = Menu.buildFromTemplate(standardMenuTemplate);
-const textInputMenu = Menu.buildFromTemplate(textMenuTemplate);
+const popupContextMenu = (event, params) => {
+  // As of Electron 4, Menu.popup no longer accepts being called with the signature popup(remote.getCurrentWindow())
+  // It must be passed as an object with the window key. Is this change silly? Yes. Will we know why it was done? No.
+  const menuPopupArgs = {
+    window: remote.getCurrentWindow()
+  };
 
-const popupContextMenu = (event) => {
-  switch (event.target.nodeName) {
-    case 'VIDEO':
-    case 'IMG':
-      if (event.target.src && event.target.src.length) {
-        let mediaType = event.target.nodeName === 'IMG' ? 'Image' : 'Video';
+  switch (params.mediaType) {
+    case 'video':
+    case 'image':
+      if (params.srcURL && params.srcURL.length) {
+        let mediaType = params.mediaType[0].toUpperCase() + params.mediaType.slice(1);
         const mediaInputMenu = Menu.buildFromTemplate([{
           label: `Save ${mediaType} As...`,
           click: () => {
             // This call *would* do this in one line, but is only a thing in IE (???)
-            // document.execCommand('SaveAs', true, event.target.src);
+            // document.execCommand('SaveAs', true, params.srcURL);
             const link = document.createElement('a');
-            link.href = event.target.src;
-            // Leaving the URL root results in the file extension being truncated.
-            // The resulting filename from this also appears to be consistent with
-            // saving the image via dragging or the Chrome context menu...winning!
-            link.download = event.target.src.replace('blob:https://messages.android.com/', '');
+            link.href = params.srcURL;
+            /*
+             * Leaving the URL root results in the file extension being truncated.
+             * The resulting filename from this also appears to be consistent with
+             * saving the image via dragging or the Chrome context menu...winning!
+             *
+             * Since the URL change from messages.android.com, the URL root of the files
+             * is messages.google.com (note the lack of /web/ in the path)
+             */
+            link.download = params.srcURL.replace('blob:https://messages.google.com/', '');
+            // Trigger save dialog by clicking the "link"
             document.body.appendChild(link);
             link.click();
             document.body.removeChild(link);
@@ -86,12 +95,53 @@ const popupContextMenu = (event) => {
       }
       break;
     default:
-      if (event.target.isContentEditable) {
-        textInputMenu.popup(remote.getCurrentWindow());
+      if (params.isEditable) {
+        const textMenuTemplateCopy = [...textMenuTemplate];
+        if (window.spellCheckHandler && params.misspelledWord && typeof params.misspelledWord === 'string') {
+          const booboo = params.selectionText;
+          textMenuTemplateCopy.unshift({
+            type: 'separator'
+          });
+          textMenuTemplateCopy.unshift({
+            label: `Add ${booboo} to Dictionary`,
+            click: () => {
+              // Immediately clear red underline
+              event.sender.replaceMisspelling(booboo);
+              // Add new custom word to dictionary for the current session
+              window.spellCheckHandler.spellCheckerTable[window.spellCheckHandler.selectedDictionary].spellChecker.addWord(booboo);
+              // Send new custom word to main process so it will be added to the dictionary at the start of future sessions
+              ipcRenderer.send(EVENT_SPELL_ADD_CUSTOM_WORD, {
+                newCustomWord: booboo
+              });
+            }
+          });
+          // Hunspell always seems to return the best choices at the end of the array, so reverse it, then limit to 8 suggestions
+          const suggestions = window.spellCheckHandler.getSuggestion(params.misspelledWord).reverse().slice(0, 8);
+          if (suggestions && suggestions.length) {
+            textMenuTemplateCopy.unshift({
+              type: 'separator'
+            });
+            suggestions.map((correction) => {
+              let item = {
+                label: correction,
+                click: () => {
+                  return event.sender.replaceMisspelling(correction);
+                }
+              };
+
+              textMenuTemplateCopy.unshift(item);
+            });
+          }
+        }
+        const textInputMenu = Menu.buildFromTemplate(textMenuTemplateCopy);
+        textInputMenu.popup(menuPopupArgs);
       } else { // Omit options pertaining to input fields if this isn't one
-        standardInputMenu.popup(remote.getCurrentWindow());
+        const standardInputMenu = Menu.buildFromTemplate(standardMenuTemplate);
+        standardInputMenu.popup(menuPopupArgs);
       }
   }
 };
 
-module.exports = popupContextMenu;
+export {
+  popupContextMenu
+};